src/app/core/security.py
from passlib.context import CryptContext
from datetime import datetime, timedelta
from jose import jwt
from app.core.config import settings

pwd_context = CryptContext(schemes=["pbkdf2_sha256"], deprecated="auto")



def hash_password(password: str) -> str:
    return pwd_context.hash(password)




def verify_password(plain_password:str, hashed_password : str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)


def create_access_token(subject:str, expires_delta: int | None = None) -> str:
    expire = datetime.utcnow() + timedelta(minutes=(expires_delta or settings.ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode = {"exp":expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode,settings.SECRET_KEY,algorithm=settings.ALGORITHM)
    return encoded_jwt


def decode_access_token(token:str):
    payload = jwt.decode(token,settings.SECRET_KEY,algorithms=[settings.ALGORITHM])
    return payload